frontend: integrate App.tsx & MainPage.tsx
