+++ frontend/src/interface/BaseArticleInterface.ts
@@ -1,0 +1,8 @@
+export default interface BaseArticleProps {
+  id: number;
+  icon: React.ReactNode;
+  title: string;
+  date: string;
+  description: string;
+  content: string;
+}
